+++ packages/react/src/components/table/table.react-template.tsx
@@ -1,8 +1,8 @@
-import { Table } from "dso-toolkit";
+import { Table, TableSorting } from "dso-toolkit";
 import * as React from "react";
 import { clsx } from "clsx";
 
-import { DsoTable } from "../../components";
+import { DsoIcon, DsoTable } from "../../components";
 import { ComponentImplementation } from "../../templates";
 
 export const reactTable: ComponentImplementation<Table<JSX.Element>> = {
@@ -10,6 +10,18 @@ export const reactTable: ComponentImplementation<Table<JSX.Element>> = {
   implementation: "react",
   template: () =>
     function tableTemplate({ noModal, content, role, verticalLines }) {
+      function getAriaRoledescription(sorting: TableSorting | undefined): string | undefined {
+        if (sorting === "ascending") {
+          return "sorteer oplopend knop";
+        }
+
+        if (sorting === "descending") {
+          return "sorteer aflopend knop";
+        }
+
+        return undefined;
+      }
+
       return (
         <DsoTable noModal={noModal}>
           <table className={clsx("table", { ["dso-table-vertical-lines"]: verticalLines })} role={role}>
@@ -17,9 +29,24 @@ export const reactTable: ComponentImplementation<Table<JSX.Element>> = {
 
             <thead>
               <tr>
-                {content.head.map((col) => (
-                  <th scope="col" key={col} dangerouslySetInnerHTML={{ __html: col }}></th>
-                ))}
+                {content.head.map((col, cellIndex) =>
+                  "sortable" in col ? (
+                    <th scope="col" key={cellIndex} aria-sort={col.sorting}>
+                      <button
+                        type="button"
+                        aria-roledescription={getAriaRoledescription(col.sorting)}
+                        className={clsx("dso-tertiary dso-sort", { "dso-sort-active": !!col.sorting })}
+                      >
+                        <span>{col.label}</span>
+                        <DsoIcon icon={col.sorting ? "sort-" + col.sorting : "sort"}></DsoIcon>
+                      </button>
+                    </th>
+                  ) : (
+                    <th scope="col" key={cellIndex}>
+                      {col.label}
+                    </th>
+                  ),
+                )}
               </tr>
             </thead>
             <tbody>